2008-04-17  Mark Rowe  <mrowe@apple.com>
        Rubber-stamped by Dan Bernstein.

        Fix the Mac build.

        * css/CSSMutableStyleDeclaration.cpp:
        (WebCore::CSSMutableStyleDeclaration::getLayeredShorthandValue): Update to use size_t in place of unsigned.
        * dom/NamedNodeMap.h: Ditto.
        * page/AccessibilityObject.h: Forward-declare AccessibilityObjectWrapper in the non-Objective-C case too.

2008-04-17  Mark Rowe  <mrowe@apple.com>

        Rubber-stamped by Dan Bernstein.

        Fix the Mac build.

        * WebView/WebFrame.mm: Define HAVE_ACCESSIBILITY before including AccessibilityObject.h and AXObjectCache.h to get things building
        for now.  This comes from config.h in WebCore but we don't have an equivalent in WebKit so we'll need to work out the correct place
        for this to live going forward.


Canonical link: https://commits.webkit.org/25503@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@31992 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/dom/NamedNodeMap.h
@@ -56,7 +56,7 @@ public:
     virtual PassRefPtr<Node> setNamedItem(Node*, ExceptionCode&) = 0;
 
     virtual PassRefPtr<Node> item(unsigned index) const = 0;
-    virtual unsigned length() const = 0;
+    virtual size_t length() const = 0;
 };
 
 } //namespace
